[Next Gen] refactor(codegen): split v-for item prop emission into item_props_merge
